backend/requirements.txt
Flask
flask-cors
passlib[bcrypt]
python-jose
mysql-connector-python
gunicorn
